\name{demo-package}
\alias{demo-package}
\alias{demo}
\docType{package}
\title{
\packageTitle{demo}
}
\description{
\packageDescription{demo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{demo}
\packageIndices{demo}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{demo}

Maintainer: \packageMaintainer{demo}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
