\name{simpleTrackAnalysis-package}
\alias{simpleTrackAnalysis-package}
\alias{simpleTrackAnalysis}
\docType{package}
\title{
\packageTitle{simpleTrackAnalysis}
}
\description{
\packageDescription{simpleTrackAnalysis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{simpleTrackAnalysis}
\packageIndices{simpleTrackAnalysis}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{simpleTrackAnalysis}

Maintainer: \packageMaintainer{simpleTrackAnalysis}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ trackAnalysis }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
